+++ packages/storybook/src/utils/utilities.ts
@@ -1,7 +1,10 @@
-import { readJson, readJsonFile, Tree } from '@nrwl/devkit';
+import { ExecutorContext, readJson, readJsonFile, Tree } from '@nrwl/devkit';
 import { CompilerOptions } from 'typescript';
 import { storybookVersion } from './versions';
-import path = require('path');
+import { StorybookConfig } from '../executors/models';
+import { constants, copyFileSync, mkdtempSync, statSync } from 'fs';
+import { tmpdir } from 'os';
+import { basename, join, sep } from 'path';
 
 export const Constants = {
   addonDependencies: ['@storybook/addons'],
@@ -128,3 +131,54 @@ export type TsConfig = {
   exclude?: string[];
   references?: Array<{ path: string }>;
 };
+
+export function findOrCreateConfig(
+  config: StorybookConfig,
+  context: ExecutorContext
+): string {
+  if (config.configFolder && statSync(config.configFolder).isDirectory()) {
+    return config.configFolder;
+  } else if (
+    statSync(config.configPath).isFile() &&
+    statSync(config.pluginPath).isFile() &&
+    statSync(config.srcRoot).isFile()
+  ) {
+    return createStorybookConfig(
+      config.configPath,
+      config.pluginPath,
+      config.srcRoot
+    );
+  } else {
+    const sourceRoot = context.workspace.projects[context.projectName].root;
+    if (statSync(join(context.root, sourceRoot, '.storybook')).isDirectory()) {
+      return join(context.root, sourceRoot, '.storybook');
+    }
+  }
+  throw new Error('No configuration settings');
+}
+
+function createStorybookConfig(
+  configPath: string,
+  pluginPath: string,
+  srcRoot: string
+): string {
+  const tmpDir = tmpdir();
+  const tmpFolder = `${tmpDir}${sep}`;
+  mkdtempSync(tmpFolder);
+  copyFileSync(
+    configPath,
+    `${tmpFolder}/${basename(configPath)}`,
+    constants.COPYFILE_EXCL
+  );
+  copyFileSync(
+    pluginPath,
+    `${tmpFolder}/${basename(pluginPath)}`,
+    constants.COPYFILE_EXCL
+  );
+  copyFileSync(
+    srcRoot,
+    `${tmpFolder}/${basename(srcRoot)}`,
+    constants.COPYFILE_EXCL
+  );
+  return tmpFolder;
+}
